ESP-Home: Communication between ESP and HomeAssistant
ESP-Home shows an example/snippet as to how data would be sent from the microcontroller, ESP32WROOM to Home Assistant via a broker on MQTT. Although, this connection has been established, it is unstable. As it constantly connects and disconnects from this broker, sending data(a random number to simulate temperature value) just few times.